src/requirements/local.txt
# Local/Development requirements
-r base.txt

# Debug
django-debug-toolbar>=4.3

# Testing
pytest>=8.0
pytest-django>=4.8
pytest-cov>=4.1
factory-boy>=3.3

# Code quality
black>=24.0
isort>=5.13
flake8>=7.0
